Setup the layout for all pages.
Created the layout with a Venue/Date block at the left and a general
info table at the right. The center part of it is the main content.
Need to fetch the Venue/Date details from the db. For now, the date
has been picked from previous year's web page.