Initial commit of the redesigned site
We have decided to go the Blueprint way and re-organised the code. We
will also be using Flask-Bootstrap extension for the site to use
bootstrap.

Now, the site has two important top level folders under accuconf -
nikola and proposals.

The 'proposals' folder contains all the code for the dynamic part of the
site - the one that handles proposal submission, review and session
scheduling.

The 'nikola' folder contains data for the static part of the site. The
contents of the 'output' folder from 'nikola build' command should be
copied into the 'static' folder of this top level folder.

Subsequently, there will be an 'admin' top level folder that has also
has dynamic code for handling admin activities like adding reviewers
and committee members.