Redesigned site with Nikola and Flask integration.
This now has both nikola and flask integrated. The "output" folder from
the nikola build command should be copied over to accuconf/nikola/static.
The code has been moved and refactored to work this approach. Nikola and
Proposals are two blueprints that serve this app. All tests are passing
again with this approach.