Do not use Path.cwd() for wsgi scripts.
Generally, the wsgi is in the web server's Doc Root and Path.cwd()
would return the path where the wsgi script is located. But the
application code itself generally resides outside of the docroot and
the code that here caculates the basedir based on the application
source code. So, added a MODULE_PATH configurable that holds the
deployment path of the accuconfweb folder.