Creating "accuconf" theme, which is based on bootstrap3-jinja
This is an initial version of the theme for accuconf. This is a three
column mode where the menu is at the left, the main content is at the
right and the sponsor logo is at the right.

The theme will continue to evolve with the conference details coming
right below the sponsor logos.