Move the sponsors to the left and use macros
Moved the sponsors block to the left in wide screens. This will now
show up right below the navigation links on wide screen and as the
bottom block on small devices.

Additionally, added macros for non-nikola managed content (like
sponsors, committee members, dates, etc.) so that it is easy to edit
all those.