Add missing cpassword to existing tests.
All the test that submitted proposals did not include the confirmation
password. Now that the code expects this field to be present, many
tests failed.

The exising tests were not really deficient because matching passwords
are checked (if I understand correctly) by javascript, but the recent
code changes make no such assumptions.